Marcus feedback: move connector checks to core to avoid duplicated code
